Expanded deleteGame so only the host is able to delete a game

+++ src/main/java/ch/uzh/ifi/hase/soprafs22/service/GameService.java
@@ -47,7 +47,14 @@ public class GameService {
         return foundGame;
     }
 
-    public void deleteGameByGameCode(Long gameCode){
+    public void deleteGameByGameCode (Long gameCode, User user)throws ResponseStatusException{
+        Session foundSession = sessionService.getSessionByGameCode(gameCode.intValue());
+        if(!user.getUsername().equals(foundSession.getHostUsername())){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Only the host can end a game");
+        }
+
+        foundSession.setHasGame(false);
+
         gameRepository.deleteByGameCode(gameCode);
     }
